requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-cors-headers==3.8.0
psycopg2-binary==2.9.1
django-cors-headers==3.8.0
django-rest-swagger==2.2.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
swagger-spec-validator==2.7.3
Markdown==3.3.4
python-dotenv==0.19.0
PyJWT==2.1.0
flake8==3.9.2
pydocstyle==6.1.1
boto3==1.18.60
Pillow==8.3.2
django-storages==1.12
SQLAlchemy==1.4.25
pandas==1.3.3
pre-commit==2.15.0
black==21.9b0
pylint==2.11.1
bcrypt==3.2.0
firebase-admin==5.1.0
pyfcm==1.5.4